refactor commandcategories to constants
