+++ src/commands/a350x/simbrief.ts
@@ -1,5 +1,7 @@
-import { CommandCategories, CommandDefinition } from '../index';
+import { CommandDefinition } from '../index';
+import { CommandCategories } from '../../constants';
 import { createEmbed } from '../../lib/embed';
+
 
 export const simbrief: CommandDefinition = {
     names: ['simbrief'],
