+++ src/commands/moderation/rules.ts
@@ -1,5 +1,6 @@
 /* eslint-disable max-len */
-import { CommandCategories, CommandDefinition } from '../index';
+import { CommandDefinition } from '../index';
+import { CommandCategories } from '../../constants';
 import { createEmbed } from '../../lib/embed';
 
 export const rules: CommandDefinition = {
